[SPARK-38753][SQL][TESTS] Move the tests for `WRITING_JOB_ABORTED` to `QueryExecutionErrorsSuite`
### What changes were proposed in this pull request?
Move test for the error class `WRITING_JOB_ABORTED` from `DataSourceV2Suite.scala` to `QueryExecutionErrorsSuite`.

### Why are the changes needed?
To improve code maintenance - all tests for error classes are placed to Query.*ErrorsSuite. Also the exception is raised from [QueryExecutionErrors](https://github.com/apache/spark/blob/073fd2ad5c16d193725954e76ce357e4a9d97449/sql/catalyst/src/main/scala/org/apache/spark/sql/errors/QueryExecutionErrors.scala#L665-L670), so, tests should be in `QueryExecutionErrorsSuite` for consistency.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
By running the affected test suites:
```
$ build/sbt "sql/test:testOnly *QueryExecutionErrorsSuite"
$ build/sbt "sql/test:testOnly *DataSourceV2Suite"
```

Closes #36196 from MaxGekk/move-tests-WRITING_JOB_ABORTED.

Authored-by: Max Gekk <max.gekk@gmail.com>
